web and desktop handle states
