+++ src/libs/actions/Session/AppleSignIn/index.ios.js
@@ -1,11 +1,8 @@
 /* eslint-disable @lwc/lwc/no-async-await */
 import _ from 'underscore';
 import {appleAuth} from '@invertase/react-native-apple-authentication';
-import ONYXKEYS from '../../../../ONYXKEYS';
-import CONST from '../../../../CONST';
-import * as Localize from '../../../Localize';
 import * as API from '../../../API';
-import DateUtils from '../../../DateUtils';
+import * as ONYX_DATA from './ONYX_DATA';
 
 async function beginAppleSignIn() {
     // performs login request
@@ -16,6 +13,7 @@ async function beginAppleSignIn() {
         // Note: it appears putting FULL_NAME first is important, see issue #293
         requestedScopes: [appleAuth.Scope.FULL_NAME, appleAuth.Scope.EMAIL],
     });
+
     console.log('finished login request', appleAuthRequestResponse);
 
     // get current authentication state for user
@@ -25,51 +23,11 @@ async function beginAppleSignIn() {
 
     // use credentialState response to ensure the user is authenticated
     const idToken = appleAuthRequestResponse.identityToken;
-    const optimisticData = [
-        {
-            onyxMethod: CONST.ONYX.METHOD.MERGE,
-            key: ONYXKEYS.ACCOUNT,
-            value: {
-                ...CONST.DEFAULT_ACCOUNT_DATA,
-                isLoading: true,
-            },
-        },
-    ];
-
-    const successData = [
-        {
-            onyxMethod: CONST.ONYX.METHOD.MERGE,
-            key: ONYXKEYS.ACCOUNT,
-            value: {
-                isLoading: false,
-            },
-        },
-        {
-            onyxMethod: CONST.ONYX.METHOD.MERGE,
-            key: ONYXKEYS.CREDENTIALS,
-            value: {
-                validateCode: null,
-            },
-        },
-    ];
-
-    const failureData = [
-        {
-            onyxMethod: CONST.ONYX.METHOD.MERGE,
-            key: ONYXKEYS.ACCOUNT,
-            value: {
-                isLoading: false,
-                errors: {
-                    [DateUtils.getMicroseconds()]: Localize.translateLocal('loginForm.cannotGetAccountDetails'),
-                },
-            },
-        },
-    ];
 
     if (credentialState === appleAuth.State.AUTHORIZED) {
         // user is authenticated
         console.log('making API request');
-        const result = await API.makeRequestWithSideEffects('AuthenticateApple', {idToken}, {optimisticData, successData, failureData});
+        const result = await API.makeRequestWithSideEffects('AuthenticateApple', {idToken}, ...ONYX_DATA);
         console.log('RESULT: ', result);
     } else {
         throw new Error('sign in failed');
