+++ src/pages/signin/LoginForm.js
@@ -84,6 +84,28 @@ class LoginForm extends React.Component {
             return;
         }
         this.input.focus();
+
+        const clientId = 'com.expensify.expensifylite.AppleSignIn';
+        const redirectURI = 'https://www.expensify.com/partners/apple/loginCallback';
+        const scope = 'name email';
+        const state = '';
+        const script = document.createElement('script');
+        script.src = 'https://appleid.cdn-apple.com/appleauth/static/jsapi/appleid/1/en_US/appleid.auth.js';
+        script.async = true;
+        script.onload = () => {
+            window.AppleID.auth.init({
+                clientId,
+                scope,
+                redirectURI,
+                state,
+                usePopup: false,
+            });
+        };
+
+        document.body.appendChild(script);
+        this.cleanupScript = () => {
+            document.body.removeChild(script);
+        };
     }
 
     componentDidUpdate(prevProps) {
@@ -94,6 +116,13 @@ class LoginForm extends React.Component {
             return;
         }
         this.input.focus();
+    }
+
+    componentWillUnmount() {
+        if (!this.cleanupScript) {
+            return;
+        }
+        this.cleanupScript();
     }
 
     /**
@@ -196,7 +225,7 @@ class LoginForm extends React.Component {
                     <DotIndicatorMessage style={[styles.mv2]} type="success" messages={{0: this.props.closeAccount.success}} />
                 )}
                 { // We need to unmount the submit button when the component is not visible so that the Enter button
-                  // key handler gets unsubscribed and does not conflict with the Password Form
+                    // key handler gets unsubscribed and does not conflict with the Password Form
                     this.props.isVisible && (
                         <View style={[styles.mt5]}>
                             <FormAlertWithSubmitButton
